requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
coverage==6.4.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-debug-toolbar==3.6.0
django-debug-toolbar-force==0.1.8
django-filter==22.1
django-nine==0.2.5
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
urllib3==1.26.12
